Migrate the library to Qiskit 2.0 (#973)
* Auto Commit

* Auto Commit tests

* Making lint, style changes with some migration fixes

* Fix copyright

* Fix copyright

* Fix copyright

* Fix copyright

* Fix copyright

* Fix copyright

* 'QuantumCircuit' objects implement '__hash__' deterministically

* Fix lint

* Lower scipy version requirement

* Relax scipy version

* Fix mypy

* Update `qiskit.primitives.StatevectorSampler`

* Update `qiskit.primitives.StatevectorSampler`

* Fix adjust_num_qubits.py

* Fix mypy and lint

* Partially update type hinting

* Update copyright and spelling

* Implement circuit hashing.

* Fix make checks.

* Fix various numerical tests.

* Fix neural networks tests. Options are outstanding.

* Fix neural networks tests + CI checks.

* Fix copyright

* Fix for samplerv2 related unittests

* Fixed unit tests for SamplerV2

* Patch mismatching parameter in the gradients lincomb (to be revisited).

* Updated type hinting. Mypy/Lint tests to be fixed.

* Updated type hinting (pt2). Mypy/Lint tests to be fixed.

* Updated copyright

* Fixed kernel tests for >2.0

* Fixed v1 to v2 arg issue for test_spsa.

* Fix None feature-map case. Update lint, copyright, spelling, and formatting.

* Silence some mypy checks in primitives.

* Consolidate feature map requirements in fidelity base kernel with documentation.

* Deprecated RawFeatureVector

* Deprecated QNNCircuit

* Format with black

* Attempt to fix html docs

* Fix exception handling

* Update QMLSampler docs

* Add primitives in API docs

* Format with black

* Fix headers

* Update 'QNNCircuit'->'qnn_circuit' in descriptions

* Set exact primitives on TorchConnector tests

* Set exact primitives on TorchConnector tests

* Fix formatting

* Relaxation of hybrid_batch_gradient test due to stochasticity from estimator

* Added QMLEstimator

* Add tests for QML Primitives

* Fix formatting

* Add docstrings and formatting

* Fix spell

* Fix mypy

* Consolidate QMLEstimator as Estimator in unit tests

* Fix mypy issue by strong typing and casting QMLSampler aux classes

* Add documentation about QML primitives

* Add reno and fix spelling

* Fix spelling in reno

* Fix spelling

* Fix precomputed case and unit tests for Kernel based methods

* Cleaning AI slop from code assistant

* Removed artefacts from V1

* mypy and style

* mypy and style

* Fix spelling and format

* Update renos

* Update Migration-to-Qiskit-2-2695587f9e23a712.yaml

---------

+++ qiskit_machine_learning/neural_networks/neural_network.py
@@ -1,6 +1,6 @@
 # This code is part of a Qiskit project.
 #
-# (C) Copyright IBM 2020, 2024.
+# (C) Copyright IBM 2020, 2025.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -124,7 +124,7 @@ class NeuralNetwork(ABC):
     ) -> tuple[np.ndarray | None, tuple[int, ...] | None]:
         if input_data is None:
             return None, None
-        input_ = np.array(input_data)
+        input_ = np.array(input_data, dtype=float)
         shape = input_.shape
         if len(shape) == 0:
             # there's a single value in the input.
@@ -177,7 +177,7 @@ class NeuralNetwork(ABC):
     ) -> np.ndarray | None:
         if weights is None:
             return None
-        weights_ = np.array(weights)
+        weights_ = np.array(weights, dtype=float)
         return weights_.reshape(self._num_weights)
 
     def _validate_forward_output(
